Remorph Source Code/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python-headless
scikit-image
scikit-learn
torch
torchvision
facenet-pytorch
python-multipart
psutil